+++ mooncake-store/tests/e2e/CMakeLists.txt
@@ -1,4 +1,9 @@
-add_executable(clientctl clientctl.cpp)
+set(MOONCAKE_E2E_TEST_SOURCES
+    process_handler.cpp
+    client_wrapper.cpp
+)
+
+add_executable(clientctl clientctl.cpp ${MOONCAKE_E2E_TEST_SOURCES})
 target_link_libraries(clientctl PUBLIC
     mooncake_store
     cachelib_memory_allocator
@@ -7,7 +12,7 @@ target_link_libraries(clientctl PUBLIC
     ${ETCD_WRAPPER_LIB}
 )
 
-add_executable(e2e_rand_test e2e_rand_test.cpp)
+add_executable(e2e_rand_test e2e_rand_test.cpp ${MOONCAKE_E2E_TEST_SOURCES})
 target_link_libraries(e2e_rand_test PUBLIC
     mooncake_store
     cachelib_memory_allocator
@@ -17,7 +22,7 @@ target_link_libraries(e2e_rand_test PUBLIC
     ${ETCD_WRAPPER_LIB}
 )
 
-add_executable(chaosctl chaosctl.cpp)
+add_executable(chaosctl chaosctl.cpp ${MOONCAKE_E2E_TEST_SOURCES})
 target_link_libraries(chaosctl PUBLIC
     mooncake_store
     cachelib_memory_allocator
@@ -26,7 +31,7 @@ target_link_libraries(chaosctl PUBLIC
     ${ETCD_WRAPPER_LIB}
 )
 
-add_executable(client_runner client_runner.cpp)
+add_executable(client_runner client_runner.cpp ${MOONCAKE_E2E_TEST_SOURCES})
 target_link_libraries(client_runner PUBLIC
     mooncake_store
     cachelib_memory_allocator
@@ -35,7 +40,7 @@ target_link_libraries(client_runner PUBLIC
     ${ETCD_WRAPPER_LIB}
 )
 
-add_executable(chaos_test chaos_test.cpp)
+add_executable(chaos_test chaos_test.cpp ${MOONCAKE_E2E_TEST_SOURCES})
 target_link_libraries(chaos_test PUBLIC
     mooncake_store
     cachelib_memory_allocator
@@ -45,7 +50,7 @@ target_link_libraries(chaos_test PUBLIC
     ${ETCD_WRAPPER_LIB}
 )
 
-add_executable(chaos_rand_test chaos_rand_test.cpp)
+add_executable(chaos_rand_test chaos_rand_test.cpp ${MOONCAKE_E2E_TEST_SOURCES})
 target_link_libraries(chaos_rand_test PUBLIC
     mooncake_store
     cachelib_memory_allocator
